Pulled from main, adjusted project structure to include PricingPalApp.kt, which holds the logic of the app's navigation functionality and app bar.
