+++ app/src/main/java/com/example/pricingpal/view/PricingPalScaffold.kt
@@ -5,13 +5,14 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.Color
 import androidx.navigation.compose.rememberNavController
+import com.example.pricingpal.PricingPalAppBar
 import com.example.pricingpal.model.Category
 import com.example.pricingpal.model.Navigation
-import com.example.pricingpal.model.PricingPalAppBar
+
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun PricingPalScaffo(categories: HashMap<String, Category>) {
+fun PricingPalScaffold(categories: HashMap<String, Category>) {
     //Initialize navController
     val navController = rememberNavController()
 
@@ -21,7 +22,7 @@ fun PricingPalScaffo(categories: HashMap<String, Category>) {
             // Get the name of the current screen
             val currentScreen = Screen.CategoryList
             PricingPalAppBar(
-                canNavigateBack = currentScreen != Screen.CategoryList,
+                returnToCategoryPage = currentScreen != Screen.CategoryList,
                 navController = navController
             )
         },
